Plumb info call to get server information.
