+++ packages/a2/bgl/src/mcp/connector-tools.ts
@@ -2,10 +2,11 @@
  * @fileoverview The tools export for the connector.
  */
 
-import type { ConnectorInfo, ListToolResult } from "../a2/connector-manager";
+import type { ListToolResult } from "../a2/connector-manager";
 import { createTools } from "../a2/connector-manager";
+import { rpc } from "../a2/rpc";
 import { ok } from "../a2/utils";
-import { McpClient } from "./mcp-client";
+import type { CallToolContent, ListToolsTool } from "./types";
 
 export { invoke as default, describe };
 
@@ -13,27 +14,19 @@ type Configuration = {
   endpoint: string;
 };
 
-async function getClient(
-  id: string,
-  info: ConnectorInfo<Configuration>
-): Promise<Outcome<McpClient>> {
-  const url = info.configuration.endpoint;
-  const client = new McpClient(id, url);
-  const connecting = await client.connect();
-  if (!ok(connecting)) return connecting;
-  return client;
-}
-
 const { invoke, describe } = createTools<Configuration>({
   title: "MCP Server",
-  list: async (id, info) => {
-    const client = await getClient(id, info);
-    if (!ok(client)) return client;
-    const listing = await client.listTools();
-    if (!ok(listing)) return listing;
-
+  list: async (_id, info) => {
+    const listingTools = await rpc<ListToolsTool[]>({
+      path: "/mnt/mcp/call/listTools",
+      data: {
+        url: info.configuration.endpoint,
+        clientName: "Breadboard",
+      },
+    });
+    if (!ok(listingTools)) return listingTools;
     // Transform to the ToolManager format.
-    const list = listing.map<ListToolResult>((item) => {
+    const list = listingTools.map<ListToolResult>((item) => {
       return {
         url: info.url,
         description: { ...item, title: item.name },
@@ -42,13 +35,17 @@ const { invoke, describe } = createTools<Configuration>({
     });
     return { list };
   },
-  invoke: async (id, info, name, args) => {
-    const client = await getClient(id, info);
-    if (!ok(client)) return client;
-    // const connecting = await client.connect();
-    // if (!ok(connecting)) return connecting;
-    const invoking = await client.callTool(name, args);
-    if (!ok(invoking)) return invoking;
-    return { result: JSON.stringify(invoking) };
+  invoke: async (_id, info, name, args) => {
+    const callingTool = await rpc<CallToolContent[]>({
+      path: "/mnt/mcp/call/callTool",
+      data: {
+        url: info.configuration.endpoint,
+        clientName: "Breadboard",
+        name,
+        arguments: args,
+      },
+    });
+    if (!ok(callingTool)) return callingTool;
+    return { result: JSON.stringify(callingTool) };
   },
 });
